multi_agent/requirements.txt
# Multi-Agent ETF 策略分析框架依赖

# LangChain / LangGraph (Agent 编排)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM 提供商（按需选择一个）
langchain-openai>=0.2.0
# langchain-anthropic>=0.2.0

# 数据源（项目已有，列出确保安装）
akshare>=1.12.0
baostock>=0.8.9
# tushare>=1.4.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 网络搜索（可选，选一个）
duckduckgo-search>=6.0.0
langchain-community>=0.3.0
# tavily-python>=0.5.0

# 推送通知
requests>=2.31.0

# 配置
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0
